pyproject.toml
[tool.poetry]
name = "vaultsbackend"
version = "0.1.0"
description = ""
authors = ["user <sergioreklamplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.5"
djangorestframework = "^3.15.2"
web3 = "^7.7.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
